Added support for BigDecimal in GenericDao

+++ src/main/no/systema/jservices/common/dao/services/GenericDaoServiceImpl.java
@@ -4,6 +4,7 @@ import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
+import java.math.BigDecimal;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -251,7 +252,7 @@ public class GenericDaoServiceImpl<T> implements GenericDaoService<T>{
 			if (getter.startsWith("get")) {
 				returnType = method.getReturnType();
 				String field = method.getName().replace("get", "").toLowerCase();
-				if (returnType.equals(String.class) || returnType.equals(int.class) || returnType.equals(float.class) ) {
+				if (returnType.equals(String.class) || returnType.equals(int.class) || returnType.equals(BigDecimal.class) ) {
 					if (!"keys".equals(field)) {
 						createString.append(field + ",");
 						try {
@@ -358,7 +359,7 @@ public class GenericDaoServiceImpl<T> implements GenericDaoService<T>{
 			if (getter.startsWith("get")) {
 				returnType = method.getReturnType();
 				String field = method.getName().replace("get", "").toLowerCase();
-				if (returnType.equals(String.class) || returnType.equals(int.class) || returnType.equals(float.class)) {
+				if (returnType.equals(String.class) || returnType.equals(int.class) || returnType.equals(BigDecimal.class)) {
 					if (!"keys".equals(field)) {
 						updateString.append(field + " = ? ,");
 						try {
